requirements.txt
# Web: расширяется до полноценного консультационного сервиса (API, сессии, БД позже)
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
jinja2>=3.1.4
python-multipart>=0.0.12
python-dotenv>=1.0.1
httpx>=0.27.0
httpx>=0.27.2
PyYAML>=6.0.2
